Move common Pod templating logic into ceohelpers package
